Cleanup.

+++ interpreter/array.h
@@ -1,8 +1,5 @@
 /*
- * TRE interpreter
- * Copyright (c) 2005-2007,2009 Sven Klose <pixel@copei.de>
- *
- * Array-related section.
+ * tré - Copyright (c) 2005-2007,2009 Sven Klose <pixel@copei.de>
  */
 
 #ifndef TRE_ARRAY_H
@@ -11,8 +8,8 @@
 extern void trearray_init (void);
 
 extern treptr trearray_get (treptr sizes);
-extern void trearray_free (treptr);
-extern ulong trearray_get_size (treptr sizes);
+extern void   trearray_free (treptr);
+extern size_t trearray_get_size (treptr sizes);
 
 extern struct tre_sequence_type trearray_seqtype;
