+++ interpreter/builtin_sequence.h
@@ -1,8 +1,5 @@
 /*
- * TRE interpreter
- * Copyright (c) 2005-2008 Sven Klose <pixel@copei.de>
- *
- * Sequence-related section.
+ * tré - Copyright (c) 2005-2008 Sven Klose <pixel@copei.de>
  */
 
 #ifndef TRE_SEQUENCE_H
@@ -13,9 +10,9 @@
  * tre_sequence_type must be defined.
  */
 struct tre_sequence_type {
-     void (*set) (treptr, ulong, treptr);
-     treptr (*get) (treptr, ulong);
-     ulong (*length) (treptr);
+     void    (*set) (treptr, ulong, treptr);
+     treptr  (*get) (treptr, ulong);
+     size_t  (*length) (treptr);
 };
 
 extern treptr tresequence_builtin_elt (treptr);
